// src/app/models/invoice/invoice-search-criteria-view.ts
/**
 * Fee
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * InvoiceSearchCriteriaView
 */
export interface InvoiceSearchCriteriaView { 
    /**
     * Gets or Sets CancelledReason
     */
    cancelledReason?: string;
    /**
     * Gets or Sets Ids
     */
    ids?: string[];
}