// src/app/service/file-management/model/zip-view-model.ts
/**
 * File API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * ZipViewModel
 */
export interface ZipViewModel { 
    /**
     * FileNames
     */
    fileNames?: Array<string>;
    /**
     * ReturnFileName
     */
    returnFileName: string;
}